fix stack viewport
