+++ packages/core/src/RenderingEngine/VolumeViewport.ts
@@ -5,6 +5,8 @@ import ViewportType from '../enums/ViewportType';
 import Viewport from './Viewport';
 import { createVolumeActor } from './helpers';
 import { loadVolume } from '../volumeLoader';
+import vtkMath from '@kitware/vtk.js/Common/Core/Math';
+import vtkPlane from '@kitware/vtk.js/Common/DataModel/Plane';
 import vtkSlabCamera from './vtkClasses/vtkSlabCamera';
 import type { vtkSlabCamera as vtkSlabCameraType } from './vtkClasses/vtkSlabCamera';
 import { getShouldUseCPURendering } from '../init';
@@ -259,7 +261,46 @@ class VolumeViewport extends Viewport implements IVolumeViewport {
    * Reset the camera for the volume viewport
    */
   public resetCamera(resetPan = true, resetZoom = true): boolean {
-    return super.resetCamera(resetPan, resetZoom);
+    const success = super.resetCamera(resetPan, resetZoom);
+
+    const activeCamera = this.getVtkActiveCamera();
+    activeCamera.setClippingRange(0.01, activeCamera.getDistance() * 2);
+
+    const viewPlaneNormal = <Point3>activeCamera.getViewPlaneNormal();
+    const focalPoint = <Point3>activeCamera.getFocalPoint();
+    const actors = this.getActors();
+    actors.forEach((actor) => {
+      const clipPlane1 = vtkPlane.newInstance();
+      const clipPlane2 = vtkPlane.newInstance();
+
+      const scaledDistance = <Point3>[
+        viewPlaneNormal[0],
+        viewPlaneNormal[1],
+        viewPlaneNormal[2],
+      ];
+      let slabThickness = 0.1;
+      if (actor.slabThickness) {
+        slabThickness = actor.slabThickness;
+      }
+      vtkMath.multiplyScalar(scaledDistance, slabThickness);
+      // we assume that the first two clipping plane of the mapper are always
+      // the 'camera' clipping
+      clipPlane1.setNormal(viewPlaneNormal);
+      const newOrigin1 = <Point3>[0, 0, 0];
+      vtkMath.subtract(focalPoint, scaledDistance, newOrigin1);
+      clipPlane1.setOrigin(newOrigin1);
+
+      clipPlane2.setNormal(vtkMath.multiplyScalar(viewPlaneNormal, -1));
+      const newOrigin2 = <Point3>[0, 0, 0];
+      vtkMath.add(focalPoint, scaledDistance, newOrigin2);
+      clipPlane2.setOrigin(newOrigin2);
+
+      const mapper = actor.volumeActor.getMapper();
+      mapper.addClippingPlane(clipPlane1);
+      mapper.addClippingPlane(clipPlane2);
+    });
+
+    return success;
   }
 
   public getFrameOfReferenceUID = (): string => {
